Make New return a non-nil Set always.
Prior to this commit, New returns nil when given an empty argument list.  This
can be confusing in cases where you want to construct a map from an existing
slice, which may be empty, to pass as an argument.

Of course you can always make(stringset.Set) to get a non-nil Set, but that's
clumsy and it is not obvious to the caller from the API when this would be
necessary.

So instead: Change New to always return a non-nil Set, even if there are no
keys to be added. Tests updated.