Rework FromKeys to avoid constructing a slice where possible.
Rather than unpacking all the strings in a map into a slice, and rather than
packing a single string into a slice, construct the output map directly and get
rid of the slice builder. Update the tests to include a Keyer test.

This also indirectly fixes a bug in the comment for FromKeys, which asserted it
returned nil for v not of the supported types. The change to make New return a
non-nil map for empty made the comment incorrect; it is now correct again as
intended.