Make the unit tests not depend directly on the type of the set element.
Instead of baking values into each test, provide an ordered set of probe values
that are shared among all the test cases, and rewrite the cases in terms of
their offsets into the order. This ensures we can have tests for any type that
can be generated, provided the caller gives us suitable values.